+++ src/components/organisms/Header/SiteName.tsx
@@ -4,8 +4,7 @@ import Link from 'next/link'
 
 import { SITE_NAME } from '@/lib/constants'
 
-import { useShouldShowSubtitle } from './hooks/useShouldShowSubtitle'
-import { useShouldShowTrpFrog } from './hooks/useShouldShowTrpFrog'
+import { useHeaderStatus } from './hooks/useHeaderStatus'
 import styles from './index.module.scss'
 import { TitleWithPageName } from './TitleWithPageName'
 
@@ -15,15 +14,16 @@ export type TitleProps = {
 }
 
 export const SiteName = (props: TitleProps) => {
-  const shouldShowSubtitle = useShouldShowSubtitle()
-  const shouldShowTrpFrog = useShouldShowTrpFrog()
-
+  const headerStatus = useHeaderStatus()
   return (
-    <div className={styles.site_logo} data-show-icon={shouldShowTrpFrog}>
+    <div
+      className={styles.site_logo}
+      data-show-icon={headerStatus.visibleTrpFrog}
+    >
       <div className={styles.trpfrog_icon} />
       <div className={styles.site_name_wrapper}>
         <h1 className={styles.site_name}>
-          {shouldShowSubtitle ? (
+          {headerStatus.visibleSubtitle ? (
             <Link
               href="/"
               style={{ cursor: 'pointer' }}
